Stop Flags from Spearing All the Things
removed "CAN COLLIDE" parameter from createVehicles calls on flag creation. Since we also make flags invulnerable this is likely the cause of random explosions on base creation.

Using "NONE" will allow the flags to use the engine to find the nearest valid location that DOES NOT collide with another object.