creates proper armored helmet array, splits items from initVar
created separate functions for the config sort, and items split by equipment/items/vehicles.

'helmets' array was really sloppy, and I needed it for something else, so I renamed it to fit the convention of my config search. I created a new array called 'armoredHelmet' that replaces the arrays use and is calculated the same way.