Jeroen Nots Debug-stuff
This commit adds the first version of the "Ultra Wide Extended Debug Console With Extra Save Buttons!", developed by Jeroen Not.
What needs testing/is missing is to make sure that still only an Admin has the debug console once this is enabled.
Also, currently this code REQUIRES CBA but it's automatically disabled if CBA is not loaded.

Co-Authored-By: Jeroen-Notenbomer <jeroen-notenbomer@users.noreply.github.com>