Update rhs_afrf_teamLeader.sqf
Replaced the AK-104 with an AK-74M. Changed some of the primary weapon attachments. Removed the launcher. Replaced the 6P9 pistol with an MP-443 Grach. Changed uniform from green Gorka suit to EMR Summer. Changed vest to a 6B23 6sh92 spetsnaz vest with a headset. Placed gun ammunition in the vest and changed number of RGOs to 2. Replaced the Altyn helmet with a 6B27 EMR-Summer helmet. Organized the comments.