Made changes to AFRF Medic PvP loadout, fixed AFRF TL PvP loadout
Forgot the PK-AS sight for the AFRF TL, added it in with this commit.  Replaced the AK-105 with an AK-74M. Changed some of the primary weapon attachments. Replaced the 6P9 pistol with an MP-443 Grach. Changed uniform from green Gorka suit to EMR Summer. Changed vest to a 6B23 6sh92 spetsnaz vest with a headset. Placed correct gun ammunition in the vest Replaced the beanie with a 6B27 EMR-Summer helmet. Organized the comments.