Changed AFRF Marksman PvP loadout, fixed a comment in AFRF Medic PvP loadout file
Replaced the SVDM with an SVDS. Replaced the 6P9 pistol with an MP-443 Grach. Changed uniform from green Gorka suit to EMR Summer. Changed vest to a 6B23 6sh92 spetsnaz vest with a headset. Increased gun ammunition in the vest. Added a PDU-4 rangefinder. Organized the comments. Fixed one of the comments in the medic loadout being out of line