Changed RHS AFRF MG PvP loadout
Removed the launcher. Replaced the 6P9 pistol with an MP-443 Grach. Changed uniform from green Gorka suit to EMR Summer. Changed vest to a 6B23 6sh92 spetsnaz vest with a headset. Placed the same ammunition found in the gun in the vest. Replaced the Altyn helmet with a 6B27 EMR-Summer helmet. Organized the comments.