Fixed formatting for _basicMedicalSupplies and _basicMiscItems calls in loadouts. Organized indentations in both USAF and AFRF loadouts for user-friendly review.
The formatting for how _basicMedicalSupplies and _basicMiscItems  were called was broken. It has been fixed to match PR #956 on the A3-Antistasi repo. Also, I've removed unnecessary brackets in the facewear section of the USAF loadouts, and I've added formatting to the backpacks of USAF loadouts without backpacks just so that anyone wanting to make some changes can do it easily. I have also rearranged