Created GREF PvP Machine Gunner file
Set the primary weapon as an M294 PIP with some accessories. Set pistol as the PM Makarov, and the uniform as an M88 TTsKO (Mountain). Set the vest as a 6b23 TTsKO Mountain (Rifleman) with its ammo, grenades, and the usual. Set the helmet to the Kaska 6B27M TTsKO (Mountain).