Changed BAF TeamLeader PvP loadout
Set primary weapon as L119A1 (Front Grip) with an IR/flashlight and a holographic sight. Set the handgun to the L131A1. Fixed format of _basicMedicalSupplies and _basicMiscItems. Set the vest as the Osprey, removed the vest NVG. Changed glass tint to clear, and changed rangefinder from Vector 21 to Lerca 1200. Also changed TFAR radio to the AN/PRC-152