Changed BAF Marksman PvP loadout
Set primary weapon as L129A1 with an IR/flashlight and a 12x sight. Set the handgun to the L131A1. Fixed format of _basicMedicalSupplies and _basicMiscItems. Set the vest as the Osprey, removed the vest NVG. Changed glass tint to clear, and set rangefinder to the Lerca 1200. Also changed TFAR radio to the AN/PRC-152