Changed BAF PvP Machinegunner Loadout
Set primary weapon as L110A3 with an IR/flashlight and a holographic sight. Set the handgun to the L131A1. Fixed format of _basicMedicalSupplies and _basicMiscItems. Set the vest as the Osprey, removed the vest NVG. Changed glass tint to clear, and gave a pair of binoculars. Also changed TFAR radio to the AN/PRC-152