Changed BAF AAT PvP loadout
Set primary weapon as L119A1 (Front Grip) with an IR/flashlight and a holographic sight. Set the handgun to the L131A1. Set the launcher as the M136. Fixed format of _basicMedicalSupplies and _basicMiscItems. Set the vest as the Osprey, removed the vest NVG. Set backpack to an MTP Kitbag with a MAAWS HEAT round. Changed glass tint to clear, and gave a pair of binoculars. Also changed TFAR radio to the AN/PRC-152