Changed CSAT PvP Team Leader (Arid) loadout
Changed the primary weapon from the normal Katiba to a short-barreled Katiba. Replaced the ARCO 4x scope with a collimator ACO scope. Removed the suppressor from the handgun, and replaced the combat uniform with a spec ops/recon uniform. Moved the NVG goggles out of the vest and into it's proper Goggles place. Increased amount of hand grenades to 2, and removed the backpack. Removed the selectRandom function from headgear and set the helmet to the Assassin spec ops helmet. Organized the comments' indentations.