Changed CSAT PvP Machine Gunner (Arid) loadout
Added a laser pointer for the machinegun. Removed the handgun suppressor. Replaced the combat uniform with a spec ops/recon uniform. Removed the select random function for the vests (now set to the LBV Harness). Removed the NVG goggles from the vest and placed them into the goggles section. Removed the select random function for the headgear (now set to the Assassin spec ops helmet). Removed the backpack. Organized the comments.