Changed CSAT PvP AT (Arid) loadout
Replaced the normal Katiba with a short-barreled one. Replaced the tan holosight scope with an ACO scope. Removed the suppressor from the handgun. Replaced the combat uniform with a spec ops/recon uniform. Reduced number of pistol magazines to 1. Move the NVGs from the vest to the goggles section. Removed select random function from the headgear and set it to the Assassin spec ops helmet. Removed the GPS. Organized comments.