Changed NATO PvP Machine Gunner (Arid) loadout
Changed primary weapon optic to ACO and the bipod's color to sand. Changed secondary weapon to 4-five .45 ACP. Removed handgun suppressor. Removed selectRandom function for uniforms and kept uniform as the recon fatigues. Removed selectRandom function for vests and kept vest as the Carrier Rig (Green). Removed NVGs from vest. Placed one 200rnd Box with only 100 rounds in it. Removed the selectRandom function on headgear and set the helmet to the Combat Helmet (Sand). Removed the GPS and placed NVGs in the Goggles section.