Changed NATO PvP AT (Arid) loadout
Changed launcher ammunition to the 75mm HEAT rocket. Changed secondary weapon to 4-five .45 ACP. Removed handgun suppressor. Removed selectRandom function for uniforms and kept uniform as the recon fatigues. Removed selectRandom function for vests and kept vest as the Carrier Rig (Green). Removed NVGs from vest. Increased amount of primary weapon magazines to 4. Removed the selectRandom function in the helmets and set the helmet to the Combat Helmet (Sand). Removed the GPSs and placed NVGs in the Goggles section.